+++ .travis.yml
@@ -5,6 +5,6 @@ before_install:
     - pip install pytest pytest-cov
     - pip install coveralls
 script:
-    - py.test
+    - python -m pytest src/chemkinlib --cov=src/chemkinlib
 after_success:
     - coveralls
