--- Documentation/Installation.tex
@@ -1,41 +1,0 @@
-\chapter{Installation}
-
-\section{Downloading the Repository}
-The repository is available on \textbf{github} through URL : \url{https://github.com/cs207-group11/cs207-FinalProject.git}. To download :
-\begin{itemize}
-    \item Below 'Contributors', a green button to 'Clone or download' the repository is present.
-    \item Use the 'Download ZIP' option to begin the download.
-    \item Once completed, the ZIP file will be present in 'Downloads' folder of your computer (default).
-    \item Unzip this file into a directory of your choice using WinRAR or equivalent software.
-    \item Now, follow instructions in Chapter 3 : Procedure for usage.
-\end{itemize}
-This link can also be used to clone the repository to your machine through the Mac Terminal/GitBash.
-\begin{itemize}
-    \item Enter directory of your choice using appropriate 'cd' commands.
-\end{itemize} 
-\begin{mdframed}[backgroundcolor=light-gray, roundcorner=10pt,leftmargin=1, rightmargin=5, innerleftmargin=10, innertopmargin=5,innerbottommargin=5, outerlinewidth=1, linecolor=light-gray]
-\begin{lstlisting}
-$ git clone <repository URL mentioned above>
-$ cd cs207-FinalProject
-\end{lstlisting}
-\end{mdframed} 
-You are now in the repository and can follow Chapter 3 : Procedure instructions for usage.
-\newpage
-\section{Test Suite}
-The testing is an important phase of development of any project and a variety of tests are necessary before the design is released. We have defined a test module called 'test\_chemkin.py' within this repository.
-\subsection{Types of Errors}
-\begin{itemize}
-    \item ValueError: check the user's input values for incorrect variable or parameter definitions. Ex : Negative Temperature/Molar Concentration. 
-    \item AttributeError: check if the execution of the program is in the right order. Ex : If the Rate Coefficient has been calculated before attempting to calculate the Reaction Rate.
-    \item TypeError : checks if the entered parameters are real and belong to a certain data type. Ex : Modified Arrhenius factor 'b' is real and a float.
-    \item NotImplementedError : checks if a certain feature/function has been designed or is due for implementation.
-\end{itemize}
-The above types of errors are tested for within the module 'test\_chemkin.py' for erroneous input to test the robustness of the 'chemkin.py' module.
-\subsection{Running the Test Suite}
-The test suite can be run from the \textbf{Terminal} to view the results of the testing.
-\begin{mdframed}[backgroundcolor=light-gray, roundcorner=10pt,leftmargin=1, rightmargin=5, innerleftmargin=10, innertopmargin=5,innerbottommargin=5, outerlinewidth=1, linecolor=light-gray]
-\begin{lstlisting}
-#Enter repository directory
->>> python -m pytest test_chemkin.py
-\end{lstlisting}
-\end{mdframed}
