--- Documentation/README.md
@@ -1,44 +1,0 @@
-LaTeX Template for Project Report
----------------------------------
-_**Version 2.0**_
-
-Abstracted from a Major Project Report at CSED, NIT Calicut but can be
-modified easily to use for other reports also.
-
-Released under Creative Commons Attribution license (CC-BY)
-Info: http://creativecommons.org/licenses/by/3.0/
-
-### Created by
-
-[Kartik Singhal](http://people.cse.nitc.ac.in/kartik "Kartik's Academic Profile")  
-BTech CSE Batch of 2009-13  
-NIT Calicut  
-Contact Info: kartiksinghal [AT] gmail [DOT] com
-
-It is advisable to learn the basics of LaTeX before using this template.
-A good resource to start with is http://en.wikibooks.org/wiki/LaTeX/
-
-How to Start
-============
-
-Start editing with report.tex and follow the instructions given there.
-
-Revision History
-================
-
-### Version 2
-1. Changed base template to a technical project report.
-2. Added example for inserting an image.
-3. Fixed numbering bug with references and acknowledgement pages.
-
-### Version 1
-1. Contains all required pages for a complete report.
-2. Highly customizable.
-3. Examples given for citing and inserting references and footnotes.
-4. PDF generated has clickable links from contents page, citations and URLs.
-5. Provides options to insert pdf title, author name, subject and keywords.
-
-Credits
-=======
-NITC logo: Mr Chandramohan MK, Graphic Studio, Central Library, NIT Calicut  
-Hacker logo: http://www.catb.org/hacker-emblem/
