--- Documentation/abstract.tex
@@ -1,9 +1,0 @@
-\vspace{2in}
-\begin{abstract}
-This repository is  collection of Python algorithms and classes intended for research in
-Reaction Rate quantification, including the calculation of Reaction Rate Coefficients and Progress Rate for a system of chemical reactions. The library is capable of receiving an XML file as input and extracting the chemical species and quantities of interest for further manipulation. The library is designed for flexibility, extensibility and ease of use. Its software
-design follows an object-oriented approach and its code is written on Python over an Anaconda Platform. 
-The library consists of three levels of documentation: a model document available in PDF format, comprehensive code documentation within individual Python files and a lower-level user's guide in the README file available in the repository.
-This is the model document: it gives a comprehensive overview of the library's capabilities, provides procedures
-for software execution, and includes example for ease of use.
-\end{abstract}
