--- Documentation/acknow.tex
@@ -1,22 +1,0 @@
-\cleardoublepage
-%\pagebreak
-\phantomsection
-\addcontentsline{toc}{chapter}{Acknowledgements}
-\chapter*{Acknowledgments}
-\vspace{1.0in}
-<Acknowledgements here>
-\\
-\\
-\\ 
-\\
-<Name here> \\ 
-\\
-<Month and Year here>\\
-{National Institute of Technology Calicut}\\
-\newpage
-\begin{figure}[htb]
-\centering
-\includegraphics[scale=0.3]{glider} % e.g. insert ./image for image.png in the working directory, adjust scale as necessary
-\caption{<Caption here>}
-\label{fig:label} % insert suitable label, this is used to refer to a fig from within the text as shown above
-\end{figure}
