--- Documentation/future-work.tex
@@ -1,2 +1,0 @@
-\chapter{Future Work}
-In the future, we intend to add a non-trivial feature for simplifying calculations in Chemical Kinetics.
