--- Documentation/prob-definition.tex
@@ -1,3 +1,0 @@
-\chapter{Problem Definition}
-
-<Problem Definition here>
