--- Documentation/ref.tex
@@ -1,8 +1,0 @@
-\cleardoublepage
-%\pagebreak
-\phantomsection
-\addcontentsline{toc}{chapter}{References}
-\begin{thebibliography}{99}
-
-\bibitem{citation-1-name-here}Chemical Kinetics : Wikipedia,\ \url{https://en.wikipedia.org/wiki/Chemical_kinetics}
-\end{thebibliography}
