--- Documentation/title.tex
@@ -1,46 +1,0 @@
-\begin{titlepage}
-
-\begin{center}
-
-\textup{\large {\bf C207 Final Project} \\ Report/Documentation}\\[0.2in]
-
-% Title
-\Large \textbf {Chemical Kinetics Library}\\[0.3in]
-
-       \small \emph{Submitted in fulfillment as per the\\
-        requirements of the course}
-        \vspace{.2in}
-
-       {\bf CS207 : Systems Development for Computational Science}\\[0.3in]
-
-% Submitted by
-\normalsize Submitted by \\
-\large \textbf{Group 11} \\
-\begin{table}[h]
-\centering
-\begin{tabular}{lr}\hline \\
-Team Members \\ \\ \hline
-\\
-Karan R. Motwani\\
-Hannah Sim\\ 
-Shiyu Huang\\
-Haixing Yin\\ \\ \hline 
-\end{tabular}
-\end{table}
-
-\vspace{.1in}
-Under the guidance of\\
-{\textbf{Prof. David Sondak}}\\[0.2in]
-
-\vfill
-
-% Bottom of the page
-\includegraphics[scale=0.65]{logo}\\[0.1in]
-\normalsize
-\textsc{Harvard University}\\
-Cambridge, Massachusetts -- 02138 \\
-\vspace{0.3cm}
-\textbf{Fall Semester, 2017}
-\end{center}
-
-\end{titlepage}
