+++ MANIFEST.in
@@ -1,0 +1,8 @@
+include LICENSE
+include MANIFEST.in
+include README.rst
+include requirements.txt
+include setup.py
+include CS207_Documentation.pdf
+
+recursive-include src *.py *.sqlite *.xml
