+++ README.md
@@ -1,11 +1,9 @@
 ChemKinLib [![Build Status](https://travis-ci.org/cs207-group11/cs207-FinalProject.svg?branch=master)](https://travis-ci.org/cs207-group11/cs207-FinalProject.svg?branch=master) [![Coverage Status](https://coveralls.io/repos/github/cs207-group11/cs207-FinalProject/badge.svg?branch=master)](https://coveralls.io/github/cs207-group11/cs207-FinalProject?branch=master)
 ===================
 
-### Group #11 Members: Karan R. Motwani, Shiyu Huang, Hannah Sim, and Haixing Yin.
+### Group #11: Karan R. Motwani, Shiyu Huang, Hannah Sim, and Haixing Yin.
 
-ChemKinLib: Chemical Kinetics Library
-
-ChemKinLib consists of algorithms, functions, and documentation to allow the user to interpret the kinetics of a system of chemical reaction by computing the associated reaction rates. The library currently supports:
+ChemKinLib (Chemical Kinetics Library) consists of classes and functions to allow the user to interpret the kinetics of a system of chemical reaction by computing the associated reaction rates. The library currently supports:
 
 - Elementary reactions: irreversible and reversible
 
@@ -14,7 +12,7 @@ And of these reactions, the following reaction rate coefficients are supported:
 - Arrhenius
 - Modified Arrhenius
 
-This Python library is designed to exhibit flexibility, extensibility, and ease of use. The initial input provided by the user is an XML file while the individual parameters for a specific reaction are entered during runtime. (to be updated)
+This Python library is designed to exhibit flexibility, extensibility, and ease of use. The initial input provided by the user is an `.xml` file while other individual parameters (temperature and concentrations) are entered during runtime.
 
 ## Installation (to be updated)
 1) Clone the repository into you local system.
@@ -32,9 +30,8 @@ The documentation for the library can be found [here](https://github.com/cs207-g
 
 ## Dependencies
 - numpy
+- scipy
 - pytest
-
-[Note : Anaconda's standard Python platform should suffice the dependency requirements.]
 
 ## License
 See [LICENSE](https://github.com/cs207-group11/cs207-FinalProject/blob/master/LICENSE) file distributed with ChemKinLib for more information.
