+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy>=1.13.1
+scipy>=0.19.1
+pytest
