--- setup.cfg
@@ -1,2 +1,0 @@
-[tool:pytest]
-addopts = --doctest-modules --cov-report term-missing --cov .
