+++ src/chemkinlib/__init__.py
@@ -1,0 +1,2 @@
+
+from ._version import __version__
