+++ src/chemkinlib/_version.py
@@ -1,0 +1,3 @@
+
+"""Define version number here and read it from setup.py automatically"""
+__version__ = "0.1"
