+++ src/chemkinlib/config.py
@@ -1,0 +1,6 @@
+import os
+
+# sqlite data directory.
+THIS_DIRECTORY = os.path.dirname(os.path.realpath(__file__))
+DATA_DIRECTORY = os.path.abspath(
+    os.path.join(THIS_DIRECTORY, 'data'))
