+++ src/chemkinlib/data/rxnset_long.xml
