+++ src/chemkinlib/data/unrecognized_rxn.xml
